model/requirements.txt
flask 
flask-cors 
langchain 
google-generativeai 
python-dotenv 
faiss-cpu 
py-eureka-client